% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxnnet2.r
\name{cindex}
\alias{cindex}
\title{Calculate C-index of a model}
\usage{
cindex(pred, time, status)
}
\arguments{
\item{pred}{Prediction from model}

\item{time}{Event time}

\item{status}{Censoring status}
}
\description{
Calculate C-index of a model
}
\examples{
cindex()
}
