% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxnnet2.r
\name{CoxLayer}
\alias{CoxLayer}
\title{Cox output layer helper function}
\usage{
CoxLayer(input, input_eval, input_size, C, R, seed = 1, save_state = NULL)
}
\description{
Cox output layer helper function
}
\examples{
CoxLayer()
}
