% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxnnet2.r
\name{CoxNnetModel}
\alias{CoxNnetModel}
\title{Create a Cox-nnet model}
\usage{
CoxNnetModel(X, time, event, n_nodes = as.integer(ceiling(sqrt(ncol(X)))),
  dropout = 1, L2 = 0, input_dropout = 1, activation = tf$tanh,
  seed = 1, sess = NULL, save_state = NULL)
}
\arguments{
\item{X}{matrix of covariates}

\item{time}{Event time}

\item{n_nodes}{Number of hidden nodes in the hidden layer}

\item{dropout}{dropout parameter of hidden layer (probability of keeping)}

\item{L2}{Weight decay/ridge regulariztaion parameter}

\item{input_dropout}{dropout parameter for the input layer}

\item{activation}{non-linear activation function}

\item{seed}{random seed used for intializing weights}

\item{sess}{TensorFlow session.  Leave as NULL.}

\item{save_state}{list of weight parameters for restoring a model}

\item{status}{Censoring status}
}
\description{
Create a Cox-nnet model
}
\examples{
CoxNnetModel()
}
