% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxnnet2.r
\name{cross_validate_coxnnet}
\alias{cross_validate_coxnnet}
\title{Helper function to cross validate a model}
\usage{
cross_validate_coxnnet(X, time, event, nfolds = 5, CV_var = "dropout",
  var_steps = seq(0.1, 0.9, 0.1), model_params = list(),
  train_params = list(), CV_seed = 1, plot = F)
}
\arguments{
\item{X}{matrix of covariates}

\item{time}{Event time}

\item{CV_var}{model parameter to profile through cross-validation (e.g., dropout or L2)}

\item{var_steps}{sequence of values of CV_var to cross-validate}

\item{model_params}{List of additional parameters passed to CoxNnetModel}

\item{plot}{Generate a plot of the cross-validated model performance over the sequence of var_steps (requires ggplot2)}

\item{status}{Censoring status}

\item{n_folds}{Number of cross-validation folds}

\item{train_parmas}{List of additional parameters passed to train_coxnnet}

\item{cv_seed}{cross validation seed used to create folds}
}
\description{
Helper function to cross validate a model
}
\examples{
cross_validate_coxnnet()
}
