% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxnnet2.r
\name{HiddenLayer}
\alias{HiddenLayer}
\title{Hidden Llayer helper function}
\usage{
HiddenLayer(input, input_eval, input_size, output_size, dropout,
  activation = tf$tanh, seed = 1, save_state = NULL)
}
\description{
Hidden Llayer helper function
}
\examples{
HiddenLayer()
}
