% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxnnet2.r
\name{load_model_coxnnet}
\alias{load_model_coxnnet}
\title{Helper function to load a model from disk}
\usage{
load_model_coxnnet(filename, X = NULL, time = NULL, event = NULL,
  sess = NULL)
}
\arguments{
\item{filename}{Name of the file containing model}

\item{X}{Covariate matrix if the model was saved with weights_only}

\item{time}{Survival time if the model was saved with weights_only}

\item{event}{Survival event if the model was saved with weights_only}

\item{sess}{The TensorFlow session can be specified if desire}
}
\description{
Helper function to load a model from disk
}
\examples{
load_model_coxnnet()
}
