% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxnnet2.r
\name{node_output_coxnnet}
\alias{node_output_coxnnet}
\title{Calculate the node outputs from a model}
\usage{
node_output_coxnnet(model, X)
}
\arguments{
\item{model}{A trained model from which to extract the weights}

\item{X}{a matrix of data points which the node outputs should be calculated from}
}
\description{
Calculate the node outputs from a model
}
\examples{
node_output_coxnnet()
}
