% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxnnet2.r
\name{pad_coxnnet}
\alias{pad_coxnnet}
\title{Calculate partial derivatives of Cox-nnet model}
\usage{
pad_coxnnet(model, X)
}
\arguments{
\item{model}{A trained model from which to extract the weights}

\item{X}{a matrix of data points where the partial derivatives should be calculated}
}
\description{
Calculate partial derivatives of Cox-nnet model
}
\examples{
pad_coxnnet()
}
