% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxnnet2.r
\name{save_model_coxnnet}
\alias{save_model_coxnnet}
\title{Helper function in order to save a model to disk}
\usage{
save_model_coxnnet(model, filename, weights_only = T)
}
\arguments{
\item{model}{A trained model from which to extract the weights}

\item{filename}{Name of the file to save to}

\item{weights_only}{Save only model weights, or also the data.  If TRUE, data must be provided in order to reconstruct the model.}
}
\description{
Helper function in order to save a model to disk
}
\examples{
save_model_coxnnet()
}
