% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxnnet2.r
\name{train_coxnnet}
\alias{train_coxnnet}
\title{Train a Cox-nnet model}
\usage{
train_coxnnet(model, max_iters = 10000,
  optimizer = tf$train$AdamOptimizer(), eval_step = 347,
  device = "/cpu:0", verbose = F)
}
\arguments{
\item{model}{Model created from CoxNnetModel}

\item{max_iters}{Maximum number of iterations to train}

\item{optimizer}{TensorFlow optimizer used to train mdoel}

\item{eval_step}{Number of gradient descent steps before calculating cost function}

\item{device}{device to use (e.g., /cpu:0, /gpu:0)}

\item{verbose}{print out training iterations}
}
\description{
Train a Cox-nnet model
}
\examples{
train_coxnnet()
}
