% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxnnet2.r
\name{weights_coxnnet}
\alias{weights_coxnnet}
\title{Extract Cox-nnet model weights}
\usage{
weights_coxnnet(model)
}
\arguments{
\item{model}{A trained model from which to extract the weights}
}
\description{
Extract Cox-nnet model weights
}
\examples{
weights_coxnnet()
}
